This PR tries to improve the current SSE2 implementation of our filler algorithms by mimicking our SSE2 blit implementations. This setup with `LOOP_UNROLLED4` and a slightly rearranged ordering looks to improve performance by 9-20+%. Now our SSE2 implementation is officially faster than our AVX2 blit with a single color surface:

New
```
Flag: BLEND_ADD
fill: 0.019394960009958596    ( should be around the same as the other ones)
blit: 0.02273272001184523
--------------------
Flag: BLEND_SUB
fill: 0.019453979993704705   # 15% better
blit: 0.022822940023615956
--------------------
Flag: BLEND_MULT
fill: 0.04831694000167772  # 11% better
blit: 0.03571774000301957
--------------------
Flag: BLEND_MIN
fill: 0.0194999800180085  # 17% better
blit: 0.02241438002092764
--------------------
Flag: BLEND_MAX
fill: 0.01942766000283882  # 16% better
blit: 0.022537899983581154
--------------------
```

Old:
```
Flag: BLEND_ADD
fill: 0.03238646000099834  <--- this is probably a spike
blit: 0.021517279999534365
--------------------
Flag: BLEND_SUB
fill: 0.022462600001017564
blit: 0.021596119999594522
--------------------
Flag: BLEND_MULT
fill: 0.05370256000023801
blit: 0.034474519999639595
--------------------
Flag: BLEND_MIN
fill: 0.02288747999991756
blit: 0.02128549999979441
--------------------
Flag: BLEND_MAX
fill: 0.022593719998258165
blit: 0.021513199999753853
--------------------
```

Test Program:
```Py
from timeit import repeat

import pygame

pygame.init()

surf = pygame.Surface((500, 500))
surf.fill((132, 33, 200))

color = pygame.Surface((500, 500))
color.fill((24, 24, 24))

flags = [
    "BLEND_ADD",
    "BLEND_SUB",
    "BLEND_MULT",
    "BLEND_MIN",
    "BLEND_MAX",
]

G = globals()

for flag in flags:
    print(f"Flag: {flag}")
    teststr = "surf.fill((24, 24, 24), None, pygame." + flag + ")"
    l = [min(repeat(teststr, globals=G, number=1000, repeat=10)) for _ in range(5)]
    print(f"fill: {sum(l) / len(l)}")

    teststr = "surf.blit(color, (0, 0), None, pygame." + flag + ")"
    l = [min(repeat(teststr, globals=G, number=1000, repeat=10)) for _ in range(5)]
    print(f"blit: {sum(l) / len(l)}")
    print("-" * 20)
```

Some graphs:
![Figure_2](https://github.com/pygame-community/pygame-ce/assets/103119829/b73a7ec3-968b-4d1b-b3cb-f9cff0aaf146)


![Figure_1](https://github.com/pygame-community/pygame-ce/assets/103119829/1ef5713e-2d20-49c5-8060-299a21f4c83b)

Graphs made with this test program:
```Py
import pygame
import json
from timeit import repeat

pygame.init()

tests = []
for mode in [
    "ADD",
    "SUB",
    "MULT",
    "MIN",
    "MAX",
]:
    for old in ["_new"]:
        stmt = f"surf.fill(0xFF01FF51, special_flags=pygame.BLEND_{mode})"
        tests.append(("fill_sse2_" + mode + old, stmt))


setup_stmt = """
surf = pygame.Surface((size, size))
"""

MAX_SIZE = 1000
REP = 20
NUM = 1


def run_tests(tests: list):
    for filename, stmt in tests:
        data = {"title": filename, "data": []}

        for size in range(1, MAX_SIZE + 1):
            if size % 50 == 0:
                print(f"Running {filename} for size {size}...")

            g = globals()
            g["size"] = size
            times = repeat(stmt, number=NUM, repeat=REP, globals=g, setup=setup_stmt)
            data["data"].append(times)

        with open(f"{filename}.json", "w") as f:
            json.dump(data, f)
        print(f"Finished {filename}! \n \n")


run_tests(tests)
pygame.quit()
```

And visualized with this one:
```Py
import json

from matplotlib import pyplot as plt
from statistics import mean, stdev, median

plt.style.use(["dark_background", "no-latex", "notebook"])

TEST_NAME = "Fill Test"
MODE = "MIN"
LIMIT_TO_RANGE = -1
files = [
    ("fill_sse2_MULT_old", "red", MODE),
    ("fill_sse2_MULT_new", "green", MODE),
]


def plot_tests(tests: list):
    for file_name, color, mode in tests:
        try:
            with open(f"{file_name}.json", "r") as f:
                data = json.load(f)
        except FileNotFoundError:
            print(f"File {file_name}.json not found!")
            quit()

        f = mean
        match mode:
            case "MEAN":
                f = mean
            case "MIN":
                f = min
            case "MAX":
                f = max
            case "MEDIAN":
                f = median

        timings = [f(data_point) for data_point in data["data"]][:LIMIT_TO_RANGE]

        print(f"=== {file_name} ===")
        print(
            f"Total: {sum([sum(data_point) for data_point in data['data']])}\n"
            f"Mean: {mean(timings)}\n"
            f"Median: {median(timings)}\n"
            f"Stdev: {stdev(timings)}"
        )
        print()
        plt.plot(timings, color=color, label=file_name, linewidth=1)


plot_tests(files)

plt.title(TEST_NAME)
plt.xlabel("Surface size (px)")
plt.ylabel("Time (s)")
plt.legend()
plt.show()
```