@Starbuck5 2024-01-04
I don't like LOOP_UNROLLED4, because I like the prospect of keeping our binaries small, and this increases the code size a lot.

Personally, I don't think it's worth the code size increase for functions like the SSE fillers. They're already 50-100x faster than they were!